Simplify calls to mozjs StringBuffer::append(const char (&array)[ArrayLength])
In Clang/g++, type deduction figures out that both sides of the
ternary operator (see the diff) can be a char[8] array, which allows
overload resolution to choose the
StringBuffer::append(const char (&array)[ArrayLength]) overload, where
ArrayLength is a template parameter with value 8.

MSVC's type deduction cannot make this inference, however, so we need
to interchange the outer StringBuffer::append() call with the inner
ternary operator, simplifying inference. The resulting code is
arguably more robust: before this change, the code was only legal
because the ".getter" and ".setter" strings happen to have the same
length.