Fall back to switch-based interpreters in MSVC
Mozjs has two bytecode interpreters, each of which operates in
"threaded" mode* by default: one for the Javascript VM and one for the
regular expression state machine (imported from V8).

In a "threaded" interpreter, each instruction has a label; an array
stores the address of each label, indexed by the instruction opcode;
and each instruction finishes its execution by looking up the label
for the next instruction and jumping to that label with a "computed
goto."

This style of interpreter requires special && syntax (no relation to
r-value references) to get the address for each label, as well as
support for using a runtime value as the target for a goto
statement. These are supported as extensions in Clang and gcc, but a
fallback switch-based approach is necessary for MSVC, which does not
support these extensions.

Preprocessor magic makes it possible to write a single interpreter
that can use either approach, depending on a small set of macro
definitions. The V8 regex interpreter already supports both
approaches, and this patch chooses the fallback when compiling on
Windows. The mozjs Javascript interpreter supported a switch-based
fallback in a _previous_ release, but that fallback was removed,
because none of mozjs's officially supported compilers need it. This
path reverses that removal, porting the switch-based interpreter
forward to this release.

*Note: "threaded" here does not refer to multi-threaded execution. The
interpreter still runs on a single thread.