Disambiguate 'ToStringSlow' function
It is not possible to get the address of an overloaded function by its
name, because the compiler would not know which overload to
choose. This patch adds an alias for the 'ToStringSlow' overload that
needs to be added to the VMFUNCTION_LIST.

It seems that Clang and g++ do not consider the other overloads of
'ToStringSlow' (in Conversions.h), because they are template
instantiations. MSVC does consider these overloads, however, resulting
in an error when compiling the VMFUNCION_LIST without using the
disambiguating alias.