Account for mmapped memory in mozjs memory tracking
This change addresses an issue seen in the 140.3 upgrade where more
memory allocations in mozjs are made against mmapp-ed chunks of memory
rather than via malloc. We try to limit the virtual memory usage of
mozjs to within a configured limit, and this change counts mmap-ed
memory towards that limit.

Relies on implementation of oom check in mongodb scripting layer